+++ crates/bevy_pbr/src/light/mod.rs
@@ -26,6 +26,8 @@ mod point_light;
 pub use point_light::{PointLight, PointLightShadowMap};
 mod spot_light;
 pub use spot_light::SpotLight;
+mod directional_light;
+pub use directional_light::{DirectionalLight, DirectionalLightShadowMap};
 
 /// Constants for operating with the light units: lumens, and lux.
 pub mod light_consts {
@@ -81,101 +83,6 @@ pub mod light_consts {
         pub const FULL_DAYLIGHT: f32 = 20_000.;
         /// The amount of light (lux) in direct sunlight.
         pub const DIRECT_SUNLIGHT: f32 = 100_000.;
-    }
-}
-
-/// A Directional light.
-///
-/// Directional lights don't exist in reality but they are a good
-/// approximation for light sources VERY far away, like the sun or
-/// the moon.
-///
-/// The light shines along the forward direction of the entity's transform. With a default transform
-/// this would be along the negative-Z axis.
-///
-/// Valid values for `illuminance` are:
-///
-/// | Illuminance (lux) | Surfaces illuminated by                        |
-/// |-------------------|------------------------------------------------|
-/// | 0.0001            | Moonless, overcast night sky (starlight)       |
-/// | 0.002             | Moonless clear night sky with airglow          |
-/// | 0.05–0.3          | Full moon on a clear night                     |
-/// | 3.4               | Dark limit of civil twilight under a clear sky |
-/// | 20–50             | Public areas with dark surroundings            |
-/// | 50                | Family living room lights                      |
-/// | 80                | Office building hallway/toilet lighting        |
-/// | 100               | Very dark overcast day                         |
-/// | 150               | Train station platforms                        |
-/// | 320–500           | Office lighting                                |
-/// | 400               | Sunrise or sunset on a clear day.              |
-/// | 1000              | Overcast day; typical TV studio lighting       |
-/// | 10,000–25,000     | Full daylight (not direct sun)                 |
-/// | 32,000–100,000    | Direct sunlight                                |
-///
-/// Source: [Wikipedia](https://en.wikipedia.org/wiki/Lux)
-///
-/// ## Shadows
-///
-/// To enable shadows, set the `shadows_enabled` property to `true`.
-///
-/// Shadows are produced via [cascaded shadow maps](https://developer.download.nvidia.com/SDK/10.5/opengl/src/cascaded_shadow_maps/doc/cascaded_shadow_maps.pdf).
-///
-/// To modify the cascade set up, such as the number of cascades or the maximum shadow distance,
-/// change the [`CascadeShadowConfig`] component of the [`DirectionalLightBundle`].
-///
-/// To control the resolution of the shadow maps, use the [`DirectionalLightShadowMap`] resource:
-///
-/// ```
-/// # use bevy_app::prelude::*;
-/// # use bevy_pbr::DirectionalLightShadowMap;
-/// App::new()
-///     .insert_resource(DirectionalLightShadowMap { size: 2048 });
-/// ```
-#[derive(Component, Debug, Clone, Reflect)]
-#[reflect(Component, Default)]
-pub struct DirectionalLight {
-    pub color: Color,
-    /// Illuminance in lux (lumens per square meter), representing the amount of
-    /// light projected onto surfaces by this light source. Lux is used here
-    /// instead of lumens because a directional light illuminates all surfaces
-    /// more-or-less the same way (depending on the angle of incidence). Lumens
-    /// can only be specified for light sources which emit light from a specific
-    /// area.
-    pub illuminance: f32,
-    pub shadows_enabled: bool,
-    pub shadow_depth_bias: f32,
-    /// A bias applied along the direction of the fragment's surface normal. It is scaled to the
-    /// shadow map's texel size so that it is automatically adjusted to the orthographic projection.
-    pub shadow_normal_bias: f32,
-}
-
-impl Default for DirectionalLight {
-    fn default() -> Self {
-        DirectionalLight {
-            color: Color::WHITE,
-            illuminance: light_consts::lux::AMBIENT_DAYLIGHT,
-            shadows_enabled: false,
-            shadow_depth_bias: Self::DEFAULT_SHADOW_DEPTH_BIAS,
-            shadow_normal_bias: Self::DEFAULT_SHADOW_NORMAL_BIAS,
-        }
-    }
-}
-
-impl DirectionalLight {
-    pub const DEFAULT_SHADOW_DEPTH_BIAS: f32 = 0.02;
-    pub const DEFAULT_SHADOW_NORMAL_BIAS: f32 = 1.8;
-}
-
-/// Controls the resolution of [`DirectionalLight`] shadow maps.
-#[derive(Resource, Clone, Debug, Reflect)]
-#[reflect(Resource)]
-pub struct DirectionalLightShadowMap {
-    pub size: usize,
-}
-
-impl Default for DirectionalLightShadowMap {
-    fn default() -> Self {
-        Self { size: 2048 }
     }
 }
